Changed tabs to spaces, moved CSS to separate file, and cleaned up requires.
The code now uses two spaces for each indent instead of tabs. This brings it more in line with standard Ruby style guidelines.

The CSS was previously a big messy, ugly string stuck into the middle of our code. This was hard to maintain for both the CSS and the surrounding code. By moving it into a separate file, which we then simply copy into our out directory, we can maintain both much easier.

I have also deleted some of the requires from the main bin file, leaving them to the individual units to require if they need.