Output now implemented EXCEPT for e-mail replies.
Imported in the Nokigiri-based XML builder to create a formatted log from the conversations and write it to the appropriate file. It's slow, but it works. Multiple chats in the same conversation are contained inside the same file EXCEPT for email replies. These are currently ignored. Should be working soon, but will take a bit of refactoring of our handling of e-mail replies.